BUG FIX: fixing array size handling in ALLOCARGV (packed-type representation).
git-svn-id: svn+ssh://saulius.grazulis.lt/home/saulius/svn-repositories/compilers/trunk@4318 f7c8cbe2-f9e6-0310-9050-d7f47bf15973

+++ starta/interpreters/packed-type/tcodes.c
@@ -1830,8 +1830,8 @@ int JMP( INSTRUCTION_FN_ARGS )
 
 int ALLOCARGV( INSTRUCTION_FN_ARGS )
 {
-    ssize_t first = istate.ep[1].num.ssize;
-    ssize_t last = istate.ep[0].num.ssize;
+    int first = istate.ep[1].num.ssize;
+    int last = istate.ep[0].num.ssize;
     char **ptr = NULL;
     ssize_t i;
 
@@ -1844,7 +1844,7 @@ int ALLOCARGV( INSTRUCTION_FN_ARGS )
 
     if( last >= first && istate.argv != NULL ) {
 
-        ssize_t length = last - first + 1;
+        int length = last - first + 1;
 	ptr = bcalloc_array( REF_SIZE, length, 1, EXCEPTION );
 
 	BC_CHECK_PTR( ptr );
